Move field definition logic into separate file
