+++ src/components/Importer.tsx
@@ -1,85 +1,19 @@
-import React, { useMemo, useState, useEffect, useContext } from 'react';
+import React, { useMemo, useState, useEffect } from 'react';
 
 import { BaseRow } from '../parser';
 import { FileStep, FileStepState } from './file-step/FileStep';
 import { generatePreviewColumns } from './fields-step/ColumnPreview';
-import { FieldsStep, Field, FieldsStepState } from './fields-step/FieldsStep';
+import { FieldsStep, FieldsStepState } from './fields-step/FieldsStep';
 import { ProgressDisplay } from './ProgressDisplay';
-import {
-  ImporterFilePreview,
-  ImporterProps,
-  ImporterFieldProps
-} from './ImporterProps';
+import { ImporterFilePreview, ImporterProps } from './ImporterProps';
+
+// re-export from a central spot
+export { ImporterField } from './ImporterField';
+import { useFieldDefinitions } from './ImporterField';
 
 import './Importer.scss';
 import { LocaleContext } from '../locale/LocaleContext';
 import { enUS } from '../locale';
-
-// internal context for registering field definitions
-type FieldDef = Field & { instanceId: symbol };
-type FieldListSetter = (prev: FieldDef[]) => FieldDef[];
-
-const FieldDefinitionContext = React.createContext<
-  ((setter: FieldListSetter) => void) | null
->(null);
-
-// defines a field to be filled from file column during import
-export const ImporterField: React.FC<ImporterFieldProps> = ({
-  name,
-  label,
-  optional
-}) => {
-  // make unique internal ID (this is never rendered in HTML and does not affect SSR)
-  const instanceId = useMemo(() => Symbol('internal unique field ID'), []);
-  const fieldSetter = useContext(FieldDefinitionContext);
-
-  // update central list as needed
-  useEffect(() => {
-    if (!fieldSetter) {
-      console.error('importer field must be a child of importer'); // @todo
-      return;
-    }
-
-    fieldSetter((prev) => {
-      const copy = [...prev];
-      const existingIndex = copy.findIndex(
-        (item) => item.instanceId === instanceId
-      );
-
-      // add or update the field definition instance in-place
-      // (using internal field instance ID helps gracefully tolerate duplicates, renames, etc)
-      const newField = {
-        instanceId,
-        name,
-        label,
-        isOptional: !!optional
-      };
-      if (existingIndex === -1) {
-        copy.push(newField);
-      } else {
-        copy[existingIndex] = newField;
-      }
-
-      return copy;
-    });
-  }, [instanceId, fieldSetter, name, label, optional]);
-
-  // on component unmount, remove this field from list by ID
-  useEffect(() => {
-    if (!fieldSetter) {
-      console.error('importer field must be a child of importer'); // @todo
-      return;
-    }
-
-    return () => {
-      fieldSetter((prev) =>
-        prev.filter((field) => field.instanceId !== instanceId)
-      );
-    };
-  }, [instanceId, fieldSetter]);
-
-  return null;
-};
 
 export function Importer<Row extends BaseRow>({
   assumeNoHeaders,
@@ -93,7 +27,7 @@ export function Importer<Row extends BaseRow>({
   ...customPapaParseConfig
 }: ImporterProps<Row>): React.ReactElement {
   // helper to combine our displayed content and the user code that provides field definitions
-  const [fields, setFields] = useState<FieldDef[]>([]);
+  const [fields, userFieldContentWrapper] = useFieldDefinitions();
 
   const [fileState, setFileState] = useState<FileStepState | null>(null);
   const [fileAccepted, setFileAccepted] = useState<boolean>(false);
@@ -168,17 +102,15 @@ export function Importer<Row extends BaseRow>({
             }}
           />
 
-          <FieldDefinitionContext.Provider value={setFields}>
-            {
-              // render the provided child content that defines the fields
-              typeof content === 'function'
-                ? content({
-                    file: fileState && fileState.file,
-                    preview: externalPreview
-                  })
-                : content
-            }
-          </FieldDefinitionContext.Provider>
+          {userFieldContentWrapper(
+            // render the provided child content that defines the fields
+            typeof content === 'function'
+              ? content({
+                  file: fileState && fileState.file,
+                  preview: externalPreview
+                })
+              : content
+          )}
         </div>
       </LocaleContext.Provider>
     );
